api/requirements.txt
sqlalchemy
sqlalchemy-utils
databases
python-multipart
bcrypt
passlib
tortoise-orm
pyjwt
fastapi
uvicorn[standard]
gunicorn
pandas
